+++ SchoolEngine/SchoolEngine/src/SchoolEngine/core/Window.h
@@ -14,8 +14,8 @@ namespace SchoolEngine {
 		// Run before closing the window (when the application is exited)
 		void cleanUp();
 		void clear();
-		void render(SDL_Texture* tex, float sizeX, float sizeY, unsigned short  int posX, unsigned short  int posY);
-		void render(SDL_Texture* tex, float sizeX, float sizeY, unsigned short  int posX, unsigned short  int posY, float angle, SDL_Point center, SDL_RendererFlip flip);
+		void render(SDL_Texture* tex, float sizeX, float sizeY, int posX, int posY);
+		void render(SDL_Texture* tex, float sizeX, float sizeY, int posX, int posY, float angle, SDL_Point center, SDL_RendererFlip flip);
 		void display();
 	private:
 		SDL_Window* _window;
